feat: 리뷰 기능

+++ src/main/java/com/skyhorsemanpower/chatService/review/data/vo/CreateReviewRequestVo.java
@@ -5,6 +5,6 @@ import lombok.Getter;
 @Getter
 public class CreateReviewRequestVo {
     private String auctionUuid;
-    private Byte reviewRate;
+    private int reviewRate;
     private String reviewContent;
 }
